Split workflow capture review UI
