+++ apps/web/src/features/plot-workflow/PlotWorkflowPanel.tsx
@@ -1,14 +1,9 @@
-import { useEffect, useMemo, useRef, useState, type CSSProperties, type ChangeEvent, type ReactNode } from "react";
+import { useEffect, useMemo, useState, type ChangeEvent } from "react";
 
 import { fetchPlotRun } from "../../lib/api";
-import type {
-  CaptureMetadata,
-  CaptureReview,
-  HardwareStatus,
-  NormalizationCorners,
-  PlotterWorkspace,
-} from "../../types/hardware";
+import type { CaptureMetadata, HardwareStatus, NormalizationCorners, PlotterWorkspace } from "../../types/hardware";
 import type { PlotAsset, PlotRun, PlotRunSummary } from "../../types/plotting";
+import { ArtifactCard } from "./ArtifactCard";
 import {
   formatPhysicalSize,
   formatResultMeta,
@@ -20,11 +15,11 @@ import {
   getStepSummaryNote,
   isRecord,
 } from "./plotWorkflowFormatters";
+import { RunArtifactCompare } from "./RunArtifactCompare";
+import { RunCaptureReview } from "./RunCaptureReview";
 import type { PlotWorkflowController } from "./usePlotWorkflow";
 
 type PlotWorkflowMode = "workflow" | "history";
-type CaptureVariantKey = "raw" | "normalized" | "debug";
-type CornerKey = keyof NormalizationCorners;
 
 interface PlotWorkflowPanelProps {
   controller: PlotWorkflowController;
@@ -37,12 +32,6 @@ interface PlotWorkflowPanelProps {
 interface RunDetailRow {
   label: string;
   value: string;
-}
-
-interface CaptureVariantOption {
-  key: CaptureVariantKey;
-  label: string;
-  url: string;
 }
 
 function resolvePreparedArtifactUrl(publicUrl: string) {
@@ -196,58 +185,6 @@ function getPreparedPageAspectRatio(run: PlotRun | null): number | null {
   return pageWidthMm / pageHeightMm;
 }
 
-function getSelectedVariantFooter({
-  capture,
-  selectedVariant,
-  resultFooter,
-}: {
-  capture: CaptureMetadata | null;
-  selectedVariant: CaptureVariantOption | null;
-  resultFooter: string | null;
-}) {
-  if (
-    selectedVariant?.key === "normalized" &&
-    capture?.normalized?.metadata.frame?.kind !== "page_aligned"
-  ) {
-    return resultFooter ? `${selectedVariant.label} · Legacy frame · ${resultFooter}` : "Normalized · Legacy frame";
-  }
-  if (selectedVariant && resultFooter) {
-    return `${selectedVariant.label} · ${resultFooter}`;
-  }
-  if (selectedVariant) {
-    return selectedVariant.label;
-  }
-  return resultFooter;
-}
-
-function getNormalizedPageAspectRatio(capture: CaptureMetadata | null): number | null {
-  const frame = capture?.normalized?.metadata.frame;
-  if (!frame || frame.kind !== "page_aligned" || frame.page_width_mm <= 0 || frame.page_height_mm <= 0) {
-    return null;
-  }
-  return frame.page_width_mm / frame.page_height_mm;
-}
-
-function getSelectedVariantFrameStyle({
-  capture,
-  selectedVariant,
-}: {
-  capture: CaptureMetadata | null;
-  selectedVariant: CaptureVariantOption | null;
-}): CSSProperties | undefined {
-  if (selectedVariant?.key === "normalized") {
-    const normalizedAspectRatio = getNormalizedPageAspectRatio(capture);
-    if (normalizedAspectRatio !== null) {
-      return { aspectRatio: `${normalizedAspectRatio}` };
-    }
-    const legacyAspectRatio = capture?.normalized?.metadata.output.aspect_ratio;
-    if (legacyAspectRatio) {
-      return { aspectRatio: `${legacyAspectRatio}` };
-    }
-  }
-  return undefined;
-}
-
 function getCollapsedRowMeta(summary: PlotRunSummary, includeTimestamp: boolean) {
   const tokens = [formatRunKindLabel(summary.asset_kind)];
   if (summary.purpose !== "normal") {
@@ -283,421 +220,6 @@ function getResultFooter({
     return run.error ?? "No result saved";
   }
   return null;
-}
-
-function getCaptureVariantOptions(capture: CaptureMetadata | null): CaptureVariantOption[] {
-  if (!capture) {
-    return [];
-  }
-
-  const options: CaptureVariantOption[] = [
-    {
-      key: "raw",
-      label: "Raw",
-      url: capture.public_url,
-    },
-  ];
-  if (capture.normalized) {
-    options.push(
-      {
-        key: "normalized",
-        label: "Normalized",
-        url: capture.normalized.rectified_grayscale_url,
-      },
-      {
-        key: "debug",
-        label: "Debug",
-        url: capture.normalized.debug_overlay_url,
-      },
-    );
-  }
-  return options;
-}
-
-function getDefaultCaptureVariant(capture: CaptureMetadata | null): CaptureVariantKey {
-  return capture?.normalized ? "normalized" : "raw";
-}
-
-function ArtifactCard({
-  className,
-  title,
-  imageUrl,
-  alt,
-  emptyMessage,
-  footer,
-  headerActions,
-  frameStyle,
-  frameContent,
-}: {
-  className?: string;
-  title: string;
-  imageUrl?: string | null;
-  alt?: string;
-  emptyMessage: string;
-  footer?: string | null;
-  headerActions?: ReactNode;
-  frameStyle?: CSSProperties;
-  frameContent?: ReactNode;
-}) {
-  return (
-    <article className={className ? `artifact-card ${className}` : "artifact-card"}>
-      <header className="artifact-card-header">
-        <h3>{title}</h3>
-        {headerActions ? <div className="artifact-card-actions">{headerActions}</div> : null}
-      </header>
-      <div className="artifact-frame" style={frameStyle}>
-        {frameContent ? (
-          frameContent
-        ) : imageUrl ? (
-          <img src={imageUrl} alt={alt ?? title} />
-        ) : (
-          <div className="empty-state">{emptyMessage}</div>
-        )}
-      </div>
-      {footer ? <p className="artifact-footer">{footer}</p> : null}
-    </article>
-  );
-}
-
-function cornersEqual(left: NormalizationCorners, right: NormalizationCorners) {
-  return (
-    left.top_left[0] === right.top_left[0] &&
-    left.top_left[1] === right.top_left[1] &&
-    left.top_right[0] === right.top_right[0] &&
-    left.top_right[1] === right.top_right[1] &&
-    left.bottom_right[0] === right.bottom_right[0] &&
-    left.bottom_right[1] === right.bottom_right[1] &&
-    left.bottom_left[0] === right.bottom_left[0] &&
-    left.bottom_left[1] === right.bottom_left[1]
-  );
-}
-
-function cloneCorners(corners: NormalizationCorners): NormalizationCorners {
-  return {
-    top_left: [...corners.top_left] as [number, number],
-    top_right: [...corners.top_right] as [number, number],
-    bottom_right: [...corners.bottom_right] as [number, number],
-    bottom_left: [...corners.bottom_left] as [number, number],
-  };
-}
-
-function CaptureReviewEditor({
-  capture,
-  review,
-  busy,
-  onAccept,
-  onAdjust,
-  onReuseLast,
-}: {
-  capture: CaptureMetadata;
-  review: CaptureReview;
-  busy: boolean;
-  onAccept: () => Promise<void>;
-  onAdjust: (corners: NormalizationCorners) => Promise<void>;
-  onReuseLast: () => Promise<void>;
-}) {
-  const [draftCorners, setDraftCorners] = useState<NormalizationCorners>(() =>
-    cloneCorners(review.proposed_corners),
-  );
-  const [activeCorner, setActiveCorner] = useState<CornerKey | null>(null);
-  const [isAdjusting, setIsAdjusting] = useState(false);
-  const [isDirty, setIsDirty] = useState(false);
-  const svgRef = useRef<SVGSVGElement | null>(null);
-  const previousCaptureIdRef = useRef(capture.id);
-  const previousProposalSignatureRef = useRef(JSON.stringify(review.proposed_corners));
-  const proposalSignature = JSON.stringify(review.proposed_corners);
-
-  useEffect(() => {
-    const captureChanged = previousCaptureIdRef.current !== capture.id;
-    const proposalChanged = previousProposalSignatureRef.current !== proposalSignature;
-    if (captureChanged || (proposalChanged && !isAdjusting && !isDirty && activeCorner === null)) {
-      setDraftCorners(cloneCorners(review.proposed_corners));
-      setActiveCorner(null);
-      setIsDirty(false);
-    }
-    if (captureChanged) {
-      setIsAdjusting(false);
-    }
-    previousCaptureIdRef.current = capture.id;
-    previousProposalSignatureRef.current = proposalSignature;
-  }, [activeCorner, capture.id, isAdjusting, isDirty, proposalSignature, review.proposed_corners]);
-
-  useEffect(() => {
-    if (!isAdjusting || typeof document === "undefined") {
-      return undefined;
-    }
-    const previousOverflow = document.body.style.overflow;
-    document.body.style.overflow = "hidden";
-    return () => {
-      document.body.style.overflow = previousOverflow;
-    };
-  }, [isAdjusting]);
-
-  const hasAdjustment = !cornersEqual(draftCorners, review.proposed_corners);
-
-  function updateCorner(clientX: number, clientY: number, key: CornerKey) {
-    const svg = svgRef.current;
-    if (!svg) {
-      return;
-    }
-    const rect = svg.getBoundingClientRect();
-    if (rect.width <= 0 || rect.height <= 0) {
-      return;
-    }
-    const x = Math.max(0, Math.min(capture.width, ((clientX - rect.left) / rect.width) * capture.width));
-    const y = Math.max(0, Math.min(capture.height, ((clientY - rect.top) / rect.height) * capture.height));
-    setDraftCorners((current) => ({
-      ...current,
-      [key]: [Number(x.toFixed(1)), Number(y.toFixed(1))],
-    }));
-    setIsDirty(true);
-  }
-
-  function resetDraftCorners() {
-    setDraftCorners(cloneCorners(review.proposed_corners));
-    setActiveCorner(null);
-    setIsDirty(false);
-  }
-
-  function openAdjustMode() {
-    resetDraftCorners();
-    setIsAdjusting(true);
-  }
-
-  function closeAdjustMode() {
-    resetDraftCorners();
-    setIsAdjusting(false);
-  }
-
-  function renderCaptureCanvas(interactive: boolean) {
-    return (
-      <div
-        className={interactive ? "capture-review-frame capture-review-frame-modal" : "capture-review-frame"}
-      >
-        <svg
-          ref={interactive ? svgRef : null}
-          viewBox={`0 0 ${capture.width} ${capture.height}`}
-          className="capture-review-svg"
-          onPointerMove={(event) => {
-            if (interactive && activeCorner) {
-              updateCorner(event.clientX, event.clientY, activeCorner);
-            }
-          }}
-          onPointerUp={() => {
-            if (interactive) {
-              setActiveCorner(null);
-            }
-          }}
-          onPointerCancel={() => {
-            if (interactive) {
-              setActiveCorner(null);
-            }
-          }}
-          onPointerLeave={() => {
-            if (interactive) {
-              setActiveCorner(null);
-            }
-          }}
-        >
-          <image href={capture.public_url} x="0" y="0" width={capture.width} height={capture.height} />
-          <polygon
-            className="capture-review-polygon"
-            points={[
-              draftCorners.top_left.join(","),
-              draftCorners.top_right.join(","),
-              draftCorners.bottom_right.join(","),
-              draftCorners.bottom_left.join(","),
-            ].join(" ")}
-          />
-          {(
-            [
-              ["top_left", draftCorners.top_left, "TL"],
-              ["top_right", draftCorners.top_right, "TR"],
-              ["bottom_right", draftCorners.bottom_right, "BR"],
-              ["bottom_left", draftCorners.bottom_left, "BL"],
-            ] as const
-          ).map(([key, point, label]) => (
-            <g key={key}>
-              <circle
-                className={
-                  interactive
-                    ? "capture-review-handle capture-review-handle-interactive"
-                    : "capture-review-handle"
-                }
-                cx={point[0]}
-                cy={point[1]}
-                r={interactive ? 18 : 12}
-                onPointerDown={
-                  interactive
-                    ? (event) => {
-                        event.preventDefault();
-                        event.currentTarget.setPointerCapture(event.pointerId);
-                        setActiveCorner(key);
-                        updateCorner(event.clientX, event.clientY, key);
-                      }
-                    : undefined
-                }
-              />
-              <text
-                x={point[0] + (interactive ? 22 : 14)}
-                y={point[1] - (interactive ? 22 : 14)}
-                className="capture-review-label"
-              >
-                {label}
-              </text>
-            </g>
-          ))}
-        </svg>
-      </div>
-    );
-  }
-
-  return (
-    <>
-      <div className="capture-review-shell">
-        {renderCaptureCanvas(false)}
-        <div className="capture-review-meta">
-          <p className="capture-review-caption">
-            {`Detection: ${review.detector_method} · confidence ${review.detector_confidence.toFixed(2)}`}
-          </p>
-          <p className="capture-review-caption">
-            Accept when the suggested quad is already usable, or open adjust mode to fine-tune the page corners.
-          </p>
-        </div>
-        <div className="capture-review-actions">
-          <button
-            type="button"
-            className="artifact-variant-button artifact-variant-button-active"
-            disabled={busy}
-            onClick={() => void onAccept()}
-          >
-            Accept
-          </button>
-          <button
-            type="button"
-            className="artifact-variant-button"
-            disabled={busy}
-            onClick={openAdjustMode}
-          >
-            Adjust
-          </button>
-          <button
-            type="button"
-            className="artifact-variant-button"
-            disabled={busy || !review.reuse_last_available}
-            onClick={() => void onReuseLast()}
-          >
-            Reuse last
-          </button>
-        </div>
-      </div>
-      {isAdjusting ? (
-        <div className="capture-review-modal" role="dialog" aria-modal="true" aria-label="Adjust capture corners">
-          <div
-            className="capture-review-modal-backdrop"
-            onClick={() => {
-              if (!busy) {
-                closeAdjustMode();
-              }
-            }}
-          />
-          <div className="capture-review-modal-panel">
-            <header className="capture-review-modal-header">
-              <div>
-                <h3>Adjust capture corners</h3>
-                <p>{`Drag each corner onto the paper edge, then apply the adjusted quad.`}</p>
-              </div>
-              <button type="button" className="button-ghost" onClick={closeAdjustMode} disabled={busy}>
-                Close
-              </button>
-            </header>
-            <div className="capture-review-modal-stage">{renderCaptureCanvas(true)}</div>
-            <div className="capture-review-modal-footer">
-              <p className="capture-review-caption">
-                {`Detection: ${review.detector_method} · confidence ${review.detector_confidence.toFixed(2)}`}
-              </p>
-              <div className="capture-review-actions">
-                <button type="button" className="artifact-variant-button" disabled={busy || !isDirty} onClick={resetDraftCorners}>
-                  Reset
-                </button>
-                <button
-                  type="button"
-                  className="artifact-variant-button"
-                  disabled={busy || !review.reuse_last_available}
-                  onClick={() => void onReuseLast()}
-                >
-                  Reuse last
-                </button>
-                <button
-                  type="button"
-                  className="artifact-variant-button artifact-variant-button-active"
-                  disabled={busy || !hasAdjustment}
-                  onClick={() => void onAdjust(draftCorners)}
-                >
-                  Apply adjusted quad
-                </button>
-              </div>
-            </div>
-          </div>
-        </div>
-      ) : null}
-    </>
-  );
-}
-
-function RunCaptureReview({
-  run,
-  preparedImageUrl,
-  preparedPageAspectRatio,
-  reviewCapture,
-  review,
-  reviewBusy,
-  onAccept,
-  onAdjust,
-  onReuseLast,
-}: {
-  run: PlotRun;
-  preparedImageUrl: string | null;
-  preparedPageAspectRatio: number | null;
-  reviewCapture: CaptureMetadata;
-  review: CaptureReview;
-  reviewBusy: boolean;
-  onAccept: () => Promise<void>;
-  onAdjust: (corners: NormalizationCorners) => Promise<void>;
-  onReuseLast: () => Promise<void>;
-}) {
-  const preparedFrameStyle =
-    preparedPageAspectRatio !== null ? { aspectRatio: `${preparedPageAspectRatio}` } : undefined;
-
-  return (
-    <div className="run-artifact-compare">
-      <ArtifactCard
-        className="artifact-card-prepared"
-        title="Prepared"
-        imageUrl={preparedImageUrl}
-        alt={`Prepared output for run ${run.id}`}
-        emptyMessage="Prepared output unavailable."
-        footer={run.asset.name}
-        frameStyle={preparedFrameStyle}
-      />
-      <article className="artifact-card artifact-card-result">
-        <header className="artifact-card-header">
-          <h3>Review capture</h3>
-        </header>
-        <div className="artifact-frame artifact-frame-review">
-          <CaptureReviewEditor
-            capture={reviewCapture}
-            review={review}
-            busy={reviewBusy}
-            onAccept={onAccept}
-            onAdjust={onAdjust}
-            onReuseLast={onReuseLast}
-          />
-        </div>
-        <p className="artifact-footer">Normalization will continue after you confirm the page corners.</p>
-      </article>
-    </div>
-  );
 }
 
 function RunHeader({
@@ -798,101 +320,6 @@ function RunSourceInspect({
         </div>
       ) : null}
     </section>
-  );
-}
-
-function RunArtifactCompare({
-  preparedImageUrl,
-  preparedAlt,
-  preparedFooter,
-  resultCapture,
-  resultAlt,
-  resultFooter,
-  preparedEmptyMessage,
-  resultEmptyMessage,
-  preparedPageAspectRatio,
-}: {
-  preparedImageUrl: string | null;
-  preparedAlt: string;
-  preparedFooter: string | null;
-  resultCapture: CaptureMetadata | null;
-  resultAlt: string;
-  resultFooter: string | null;
-  preparedEmptyMessage: string;
-  resultEmptyMessage: string;
-  preparedPageAspectRatio: number | null;
-}) {
-  const variantOptions = useMemo(
-    () => getCaptureVariantOptions(resultCapture),
-    [resultCapture],
-  );
-  const defaultVariant = getDefaultCaptureVariant(resultCapture);
-  const [resultVariant, setResultVariant] = useState<CaptureVariantKey>(defaultVariant);
-
-  useEffect(() => {
-    setResultVariant(defaultVariant);
-  }, [defaultVariant, resultCapture?.id]);
-
-  const selectedVariant =
-    variantOptions.find((option) => option.key === resultVariant) ?? variantOptions[0] ?? null;
-  const variantFooter = getSelectedVariantFooter({
-    capture: resultCapture,
-    selectedVariant,
-    resultFooter,
-  });
-  const preparedFrameStyle =
-    preparedPageAspectRatio !== null ? { aspectRatio: `${preparedPageAspectRatio}` } : undefined;
-  const resultFrameStyle = getSelectedVariantFrameStyle({
-    capture: resultCapture,
-    selectedVariant,
-  });
-
-  return (
-    <div className="run-artifact-compare">
-      <ArtifactCard
-        className="artifact-card-prepared"
-        title="Prepared"
-        imageUrl={preparedImageUrl}
-        alt={preparedAlt}
-        emptyMessage={preparedEmptyMessage}
-        footer={preparedFooter}
-        frameStyle={preparedFrameStyle}
-      />
-      <ArtifactCard
-        className="artifact-card-result"
-        title="Result"
-        imageUrl={selectedVariant?.url ?? null}
-        alt={
-          selectedVariant
-            ? `${selectedVariant.label.toLowerCase()} ${resultAlt}`
-            : resultAlt
-        }
-        emptyMessage={resultEmptyMessage}
-        footer={variantFooter}
-        frameStyle={resultFrameStyle}
-        headerActions={
-          variantOptions.length > 1 ? (
-            <div className="artifact-variant-selector" role="group" aria-label="Result variant">
-              {variantOptions.map((option) => (
-                <button
-                  key={option.key}
-                  type="button"
-                  className={
-                    option.key === resultVariant
-                      ? "artifact-variant-button artifact-variant-button-active"
-                      : "artifact-variant-button"
-                  }
-                  aria-pressed={option.key === resultVariant}
-                  onClick={() => setResultVariant(option.key)}
-                >
-                  {option.label}
-                </button>
-              ))}
-            </div>
-          ) : null
-        }
-      />
-    </div>
   );
 }
